requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
configparser==5.0.2
cpprb==10.1.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
dm-tree==0.1.6
docker-pycreds==0.4.0
fasteners==0.16
flake8==3.9.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
gitdb==4.0.7
GitPython==3.1.17
glfw==2.1.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.32.0
gym==0.18.3
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mccabe==0.6.1
moviepy==1.0.3
msgpack==1.0.2
mujoco-py==2.0.2.13
neovim==0.3.1
numpy==1.21.1
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.0
pathtools==0.1.2
Pillow==8.2.0
proglog==0.1.9
promise==2.3
protobuf==3.17.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyglet==1.5.15
pynvim==0.4.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
seaborn==0.11.1
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.4.0
tensorflow-gan==2.0.0
tensorflow-hub==0.12.0
tensorflow-probability==0.12.0
termcolor==1.1.0
tf-agents==0.8.0
tqdm==4.61.0
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.5
wandb==0.10.31
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
